store pose to file

+++ include/Converter.cc
@@ -16,46 +16,46 @@ std::vector<cv::Mat> Converter::toDescriptorVector(const cv::Mat &Descriptors)
 
 cv::Mat Converter::toCvMat(const Eigen::Matrix<double,4,4> &m)
 {
-    cv::Mat cvMat(4,4,CV_64F);
+    cv::Mat cvMat(4,4,CV_32F);
     for(int i=0;i<4;i++)
         for(int j=0; j<4; j++)
-            cvMat.at<double>(i,j)=m(i,j);
+            cvMat.at<float>(i,j)=m(i,j);
 
     return cvMat.clone();
 }
 
 cv::Mat Converter::toCvMat(const Eigen::Matrix3d &m)
 {
-    cv::Mat cvMat(3,3,CV_64F);
+    cv::Mat cvMat(3,3,CV_32F);
     for(int i=0;i<3;i++)
         for(int j=0; j<3; j++)
-            cvMat.at<double>(i,j)=m(i,j);
+            cvMat.at<float>(i,j)=m(i,j);
 
     return cvMat.clone();
 }
 
 cv::Mat Converter::toCvMat(const Eigen::Matrix<double,3,1> &m)
 {
-    cv::Mat cvMat(3,1,CV_64F);
+    cv::Mat cvMat(3,1,CV_32F);
     for(int i=0;i<3;i++)
-            cvMat.at<double>(i)=m(i);
+        cvMat.at<float>(i)=m(i);
 
     return cvMat.clone();
 }
 
 cv::Mat Converter::toCvSE3(const Eigen::Matrix<double,3,3> &R, const Eigen::Matrix<double,3,1> &t)
 {
-    cv::Mat cvMat = cv::Mat::eye(4,4,CV_64F);
+    cv::Mat cvMat = cv::Mat::eye(4,4,CV_32F);
     for(int i=0;i<3;i++)
     {
         for(int j=0;j<3;j++)
         {
-            cvMat.at<double>(i,j)=R(i,j);
+            cvMat.at<float>(i,j)=R(i,j);
         }
     }
     for(int i=0;i<3;i++)
     {
-        cvMat.at<double>(i,3)=t(i);
+        cvMat.at<float>(i,3)=t(i);
     }
 
     return cvMat.clone();
@@ -64,7 +64,7 @@ cv::Mat Converter::toCvSE3(const Eigen::Matrix<double,3,3> &R, const Eigen::Matr
 Eigen::Matrix<double,3,1> Converter::toVector3d(const cv::Mat &cvVector)
 {
     Eigen::Matrix<double,3,1> v;
-    v << cvVector.at<double>(0), cvVector.at<double>(1), cvVector.at<double>(2);
+    v << cvVector.at<float>(0), cvVector.at<float>(1), cvVector.at<float>(2);
     return v;
 }
 
@@ -79,9 +79,9 @@ Eigen::Matrix<double,3,3> Converter::toMatrix3d(const cv::Mat &cvMat3)
 {
     Eigen::Matrix<double,3,3> M;
 
-    M << cvMat3.at<double>(0,0), cvMat3.at<double>(0,1), cvMat3.at<double>(0,2),
-         cvMat3.at<double>(1,0), cvMat3.at<double>(1,1), cvMat3.at<double>(1,2),
-         cvMat3.at<double>(2,0), cvMat3.at<double>(2,1), cvMat3.at<double>(2,2);
+    M << cvMat3.at<float>(0,0), cvMat3.at<float>(0,1), cvMat3.at<float>(0,2),
+         cvMat3.at<float>(1,0), cvMat3.at<float>(1,1), cvMat3.at<float>(1,2),
+         cvMat3.at<float>(2,0), cvMat3.at<float>(2,1), cvMat3.at<float>(2,2);
 
     return M;
 }
